File uploads and invoice downloads features

+++ controllers/admin.js
@@ -1,17 +1,34 @@
 const { validationResult } = require("express-validator");
 
+const fileHelper = require('../util/file');
+
 const Product = require('../models/product');
 
 exports.getAddProduct = (req, res, next) => {
-    res.render('admin/edit-product', {pageTitle: 'Add Product', path:'/admin/add-product', editing: false, hasError: false, isAuthenticated: req.session.isLoggedIn, errorMessage: null});
+    res.render('admin/edit-product', {pageTitle: 'Add Product', path:'/admin/add-product', editing: false, hasError: false, isAuthenticated: req.session.isLoggedIn, errorMessage: null, validationErrors: []});
 }
 
 exports.postAddProduct = (req, res, next) => {
     const title = req.body.title;
-    const imageUrl = req.body.imageUrl;
+    const image = req.file;
     const price = req.body.price;
     const description = req.body.description;
-    
+    if (!image) {
+        return res.status(422).render('admin/edit-product', {
+            pageTitle: 'Add Product', 
+            path:'/admin/products', 
+            editing: false,
+            product: {
+                title: title,
+                price: price,
+                description: description,
+            },
+            hasError: true,
+            errorMessage: 'Attached file is not an image.',
+            isAuthenticated: req.session.isLoggedIn
+        });
+    }
+    const imageUrl = image.path;
     const errors = validationResult(req);
     if (!errors.isEmpty()) {
         console.log(errors);
@@ -23,7 +40,6 @@ exports.postAddProduct = (req, res, next) => {
                 title: title,
                 price: price,
                 description: description,
-                imageUrl: imageUrl
             },
             hasError: true,
             errorMessage: errors.array()[0].msg,
@@ -100,10 +116,12 @@ exports.getEditProduct = (req, res, next) => {
                 product: product,
                 hasError: false,
                 errorMessage: null,
-                isAuthenticated: req.session.isLoggedIn
+                isAuthenticated: req.session.isLoggedIn,
+                validationErrors: []
             });
         })
         .catch(err => {
+            console.log(err);
             const error = new Error(err);
             error.httpStatusCode = 500;
             return next(error);
@@ -114,8 +132,10 @@ exports.postEditProduct = (req, res, next) => {
     const prodId = req.body.productId;
     const updatedTitle = req.body.title;
     const updatedPrice = req.body.price;
-    const updatedImageUrl = req.body.imageUrl;
+    const image = req.file;
     const updatedDescription = req.body.description;
+
+    
     const errors = validationResult(req);
     if (!errors.isEmpty()) {
         console.log(errors);
@@ -127,8 +147,7 @@ exports.postEditProduct = (req, res, next) => {
                 _id: prodId,
                 title: updatedTitle,
                 price: updatedPrice,
-                description: updatedDescription,
-                imageUrl: updatedImageUrl
+                description: updatedDescription
             },
             hasError: true,
             errorMessage: errors.array()[0].msg,
@@ -140,10 +159,13 @@ exports.postEditProduct = (req, res, next) => {
             if (product.userId.toString() !== req.user._id.toString()) {
                 return res.redirect('/');
             }
+            if (image) {
+                fileHelper.deleteFile(product.imageUrl);
+                product.imageUrl = image.path;
+            }
             product.title = updatedTitle;
             product.price = updatedPrice;
             product.description = updatedDescription;
-            product.imageUrl = updatedImageUrl
             return product.save().then(() => res.redirect('/admin/products'));      
         })
         .catch(err => {
@@ -156,7 +178,14 @@ exports.postEditProduct = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
     const productId = req.body.productId;
-    Product.deleteOne({_id: productId, userId: req.user._id})
+    Product.findById(productId)
+        .then(product => {
+            if (!product) {
+                return next(new Error('Product not found.'));
+            }
+            fileHelper.deleteFile(product.imageUrl);
+            return Product.deleteOne({_id: productId, userId: req.user._id});
+        })
         .then(() => res.redirect('/admin/products'))
         .catch(err => {
             const error = new Error(err);
